feat(resolve): adding currentFile to collect project
